separar login de vista de usuario
